Implements update/delete functionality for Speedruns controller...
Updates routing for the new controller methods.
Views now have links to everywhere (especially for admins).
Elements requiring Administrative access are properly safeguarded from crashes for anonymous guests.
Fixes typo in Sessions Helper comment bcesaue we hvae to fxi hte tpyos.

All topics related to the issue of this branch are satisfactorily resolved.
This marks the end of this branch.